refactor: move encoders to text helpers
