+++ resources/js/app/components/placeholder-list/placeholder-params.vue
@@ -52,7 +52,6 @@
 </template>
 <script>
 import { EventBus } from 'app/components/event-bus';
-import { placeholderParamsDecoded, placeholderParamsEncode } from 'app/helpers/placeholder-encoding';
 
 export default {
     name: 'PlaceholderParams',
@@ -89,9 +88,9 @@ export default {
     mounted() {
         this.$nextTick(() => {
             this.parameters = BEDITA?.placeholdersConfig?.[this.type] || {};
-            const decoded = placeholderParamsDecoded(this.value);
+            const decoded = this.$helpers.binaryToAsciiUtf8(this.value);
             if (decoded === 'undefined') {
-                this.newValue = placeholderParamsEncode('undefined');
+                this.newValue = this.$helpers.asciiToBinaryUtf8('undefined');
 
                 return;
             }
@@ -105,7 +104,7 @@ export default {
     methods: {
         changeParams() {
             this.oldValue = this.newValue || this.value;
-            this.newValue = placeholderParamsEncode(JSON.stringify(this.decodedValue));
+            this.newValue = this.$helpers.asciiToBinaryUtf8(JSON.stringify(this.decodedValue));
             EventBus.send('replace-placeholder', {
                 id: this.id,
                 field: this.field,
