+++ resources/js/app/helpers/text-helper.js
@@ -19,3 +19,27 @@ export function humanizeString(string) {
 export function upperCaseFirst(input) {
     return input.charAt(0).toUpperCase() + input.substr(1);
 }
+
+
+/**
+ * Encode string to base64 using text encoder to ensure that the string is in utf-8 format
+ * @param {string|undefined} value
+ * @returns {string} base64 encoded string
+ */
+export function asciiToBinaryUtf8(value = 'undefined') {
+    const utf8Bytes = new TextEncoder().encode(value);
+
+    return btoa(String.fromCharCode(...utf8Bytes));
+};
+
+/**
+ * Decode string from base64 to utf-8
+ * @param {string| undefined} value
+ * @returns {string} decoded string
+ */
+export function binaryToAsciiUtf8(value) {
+    const binaryString = atob(value);
+    const bytes = Uint8Array.from(binaryString, c => c.charCodeAt(0));
+
+    return new TextDecoder().decode(bytes);
+};
